Validation and Swagger docs: - Added validation constraints and exceptions handling; - Added Swagger Springfox UI for testing the app;

+++ src/main/java/com/inmost/tasktracker/service/impl/TaskServiceImpl.java
@@ -1,6 +1,8 @@
 package com.inmost.tasktracker.service.impl;
 
+import com.inmost.tasktracker.exception.DataProcessingException;
 import com.inmost.tasktracker.model.Task;
+import com.inmost.tasktracker.repository.StatusRepository;
 import com.inmost.tasktracker.repository.TaskRepository;
 import com.inmost.tasktracker.service.TaskService;
 import java.util.List;
@@ -10,15 +12,21 @@ import org.springframework.stereotype.Service;
 public class TaskServiceImpl implements TaskService {
     private static final String SORTING_TYPE_ASC = "asc";
     private static final String SORTING_TYPE_DESC = "desc";
+    private static final String VIEW = "View";
 
     private final TaskRepository taskRepository;
+    private final StatusRepository statusRepository;
 
-    public TaskServiceImpl(TaskRepository taskRepository) {
+    public TaskServiceImpl(TaskRepository taskRepository, StatusRepository statusRepository) {
         this.taskRepository = taskRepository;
+        this.statusRepository = statusRepository;
     }
 
     @Override
     public Task add(Task task) {
+        if (task.getStatus() == null) {
+            task.setStatus(statusRepository.getStatusByStatusName(VIEW));
+        }
         return taskRepository.save(task);
     }
 
@@ -26,7 +34,7 @@ public class TaskServiceImpl implements TaskService {
     public Task get(long taskId) {
         return taskRepository.findById(taskId)
                 .orElseThrow(() ->
-                        new RuntimeException("Failed to find a task with id: " + taskId));
+                        new DataProcessingException("Failed to find a task with id: " + taskId));
     }
 
     @Override
@@ -36,7 +44,7 @@ public class TaskServiceImpl implements TaskService {
         } else if (sortingType.toLowerCase().equals(SORTING_TYPE_ASC)) {
             return taskRepository.findAllByStatus_StatusNameOrderByAssignee_UserIdAsc(statusName);
         } else {
-            throw new RuntimeException("Wrong parameter for sorting type has been inputted "
+            throw new DataProcessingException("Wrong parameter for sorting type has been inputted "
                     + ". Please choose either \"asc\" for ascending order or \"desc\" for "
                     + " descending order.");
         }
@@ -49,7 +57,7 @@ public class TaskServiceImpl implements TaskService {
         } else if (sortingType.toLowerCase().equals(SORTING_TYPE_ASC)) {
             return taskRepository.getAllByAssigneeNotNullOrderByAssignee_UserIdAsc();
         } else {
-            throw new RuntimeException("Wrong parameter for sorting type has been inputted "
+            throw new DataProcessingException("Wrong parameter for sorting type has been inputted "
                     + ". Please choose either \"asc\" for ascending order or \"desc\" for "
                     + " descending order.");
         }
@@ -60,12 +68,16 @@ public class TaskServiceImpl implements TaskService {
         if (taskRepository.findById(task.getTaskId()).isPresent()) {
             return taskRepository.save(task);
         } else {
-            throw new RuntimeException("Failed to find a task with id: " + task.getTaskId());
+            throw new DataProcessingException("Failed to find a task with id: " + task.getTaskId());
         }
     }
 
     @Override
     public void delete(long taskId) {
-        taskRepository.deleteById(taskId);
+        if (taskRepository.findById(taskId).isPresent()) {
+            taskRepository.deleteById(taskId);
+        } else {
+            throw new DataProcessingException("Failed to find a task with id: " + taskId);
+        }
     }
 }
